refactor(vehicles): split file and spare-part helpers
